requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
soccerdata>=1.4.0
pandas>=2.0.0
lxml>=4.9.0
html5lib>=1.1
